adding architecture diagram
